[INFRA/CORE] Update docs, cleanup and refactor
