+++ internal/common/baseclient/baseclient.go
@@ -5,11 +5,9 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/SOMAS2020/SOMAS2020/internal/common/rules"
-	"github.com/SOMAS2020/SOMAS2020/pkg/miscutils"
-
 	"github.com/SOMAS2020/SOMAS2020/internal/common/gamestate"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/roles"
+	"github.com/SOMAS2020/SOMAS2020/internal/common/rules"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/shared"
 )
 
@@ -114,38 +112,6 @@ func (c *BaseClient) StartOfTurn() {}
 // BASE: Do not overwrite in team client.
 func (c *BaseClient) Logf(format string, a ...interface{}) {
 	log.Printf("[%v]: %v", c.id, fmt.Sprintf(format, a...))
-}
-
-// Role provides enumerated type for IIGO roles (President, Speaker and Judge)
-type Role int
-
-const (
-	President Role = iota
-	Speaker
-	Judge
-)
-
-func (r Role) String() string {
-	strs := [...]string{"President", "Speaker", "Judge"}
-	if r >= 0 && int(r) < len(strs) {
-		return strs[r]
-	}
-	return fmt.Sprintf("UNKNOWN Role '%v'", int(r))
-}
-
-// GoString implements GoStringer
-func (r Role) GoString() string {
-	return r.String()
-}
-
-// MarshalText implements TextMarshaler
-func (r Role) MarshalText() ([]byte, error) {
-	return miscutils.MarshalTextForString(r.String())
-}
-
-// MarshalJSON implements RawMessage
-func (r Role) MarshalJSON() ([]byte, error) {
-	return miscutils.MarshalJSONForString(r.String())
 }
 
 // GetVoteForRule returns the client's vote in favour of or against a rule.
